Implement storage as SingleCellExperiment object
Fairly major structural change. Now stored as SingleCellExperiment, but
with each observation being a sample rather than a cell. Most things are
stored in colData, with only the MMD distances stored in the assays.

Note to self: everything works, but the storage still feels non-optimal.
So I need to understand nesting in DataFrames better.

Also:
- merged calculate_sample_to_sample_MMDs and embed_sample_to_sample_MMDs
- tweaked the embeddings plots a tiny bit