Fix #20 by not clustering with too few samples
Main change is that when there are few samples, it doesn't make sense
to cluster them, so we give all samples one cluster SG10. I've set the
cutoff to <=10 for now.

I also addressed Reto's bug when 'sample_id' is included in annots_disc,
just by removing it and throwing a warning.